splitting out checks into their own module
